+++ opcode.hh
@@ -2,13 +2,37 @@
 
 #include <cstdint>
 
+#define DIRECTION_INSENSITIVE_OPS \
+    NOP = '.',      /* no-op */ \
+        ADD = '+',  /* add */ \
+        SUB = '-',  /* subtract */ \
+        MUL = '*',  /* multiply */ \
+        DIV = ':',  /* divide */ \
+        MOD = '%',  /* modulo */ \
+        PSI = '\'', /* push integer */ \
+        PSC = '"',  /* push character */ \
+        POP = ',',  /* pop */ \
+        EXT = '@',  /* exit */ \
+        INC = ')',  /* increment */ \
+        DEC = '(',  /* decrement */ \
+        AND = '&',  /* bitwise and */ \
+        IOR = 'r',  /* bitwise or */ \
+        XOR = 'x',  /* bitwise x-or */ \
+        NOT = '~',  /* bitwise not */ \
+        GTC = 'i',  /* getchar */ \
+        PTC = 'o',  /* putchar */ \
+        GTI = '?',  /* get integer */ \
+        PTI = '!',  /* put integer */ \
+        IDX = 'j',  /* index */ \
+        DUP = '2',  /* duplicate */ \
+        RND = '$',  /* random */ \
+        EXP = 'e',  /* exponential */ \
+        SWP = 'S',  /* swap */ \
+        GTM = 'T',  /* get time */ \
+        GDT = 'D'   /* get date */
+
+// The valid source tokens.
 enum opcode : int32_t {
-    NOP = '.',        // no-op
-    ADD = '+',        // add
-    SUB = '-',        // subtract
-    MUL = '*',        // multiply
-    DIV = ':',        // divide
-    MOD = '%',        // modulo
     BNG_NW = '^',     // branch if negative, pointing northwest
     BNG_NE = '7',     // branch if negative, pointing northeast
     BNG_E = '>',      // branch if negative, pointing east
@@ -19,28 +43,8 @@ enum opcode : int32_t {
     MIR_NS = '|',     // mirror north/south
     MIR_NESW = '/',   // mirror northeast/southwest
     MIR_NWSE = '\\',  // mirror northwest/southeast
-    PSI = '\'',       // push integer
-    PSC = '"',        // push character
-    POP = ',',        // pop
-    EXT = '@',        // exit
-    INC = ')',        // increment
-    DEC = '(',        // decrement
-    AND = '&',        // bitwise and
-    IOR = 'r',        // bitwise or
-    XOR = 'x',        // bitwise x-or
-    NOT = '~',        // bitwise not
-    GTC = 'i',        // getchar
-    PTC = 'o',        // putchar
-    GTI = '?',        // get integer
-    PTI = '!',        // put integer
-    SKP = '#',        // skip
-    IDX = 'j',        // index
-    DUP = '2',        // duplicate
-    RND = '$',        // random
-    EXP = 'e',        // exponential
-    SWP = 'S',        // swap
     THR_W = '{',      // west thread operator
     THR_E = '}',      // east thread operator
-    GTM = 'T',        // get time
-    GDT = 'D',        // get date
+    SKP = '#',        // skip
+    DIRECTION_INSENSITIVE_OPS
 };