+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file. See [Keep a
 
 ### Changed
 
+- Changed disassembler label format from a single integer to two integers separated by a dot.
 - Uses EXIT_FAILURE instead of hardcoding 1.
 
 ## [1.2.1] - 2023-03-28